pyproject.toml
[project]
name = "powerfill-py"
version = "0.1.0"
description = "Powerfill Python SDK"
readme = "README.md"
requires-python = ">=3.12"
authors = ["Arnaud Kayonga <a.kayonga@meshpower.co.rw>"]
dependencies = [
    "aiohttp>=3.13.2",
    "attr>=0.3.2",
    "httpx>=0.28.1",
    "openapi-python-client>=0.27.1",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
ci = [
    "python-semantic-release>=10.5.2",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = false
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
changelog_file = ""
exclude_commit_patterns = []
mode = "update"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = true

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]